+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Unreleased
 
+## v3.0.0 (2024-12-09)
+
+* Drop support for PHP < 8.2
+
 ## v2.1.0 (2024-09-25)
 
 * Support PHP 8.3